feat: Add drop down for filtering devices by pool
When there are a large number of devices available, the main devices
page can get very cluttered making it difficult to find available
devices of the type the user is looking for.

This patch adds a dropdown to the main devices page for selecting a pool
to filter the page by. The devices are collected by pool using a new
endpoint /devices/pool/<pool_id>.

Change-Id: Ic2ba5d42264b9cbf6c3836005ee7c2ee461d4fbb