feat - Change device pool filtering to query string
The endpoint for returning a subset of the devices filtered by pool was
'devices/pool/<pool_id>'. After reading the documentation of rocket some
more it seems like the better approach is to use query string for
filtering data like this.

This change switches to using an optional query string approach for
filtering the devices. The new url would look like 'devices?pool_id=1'.
This approach allowed for a unification of the devices and pool filtered
devices endpoints.

Change-Id: I0363d1f564fcb931c838ccd38da3ea517ab2462a