Update rocket from 0.4 to 0.5-rc2
This patch updates the rocket dependency from 0.4 to 0.5-rc2. Two tests
also had to be slightly modified in order to work with the new version.

The two slightly reworked tests are:
* test_html_add_devices_bad_url
* test_html_edit_devices_bad_url

This is because 0.5-rc2 added extra validation which these tests now
need to account for.

Behavior/Functionality should remain the same.

Change-Id: I3c335805746b70dea3b192bdce2b5088ddb1832a